backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Filip <mularskif@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
django-cors-headers = "^4.0.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.5.1"
faker = "^18.10.1"
drf-spectacular = "^0.26.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
